backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
python-dotenv
alembic
pytest
httpx
PyJWT
passlib[bcrypt]